fixed multiplayer camera bug where everyone would see thru 1 player's POV
instead of every player (and drone) having its own camera, there's ONE maincamera in the scene that we can tell to follow any object we want.
that object will be the player when the game starts, but can be set to the drone when a button is pressed.

the maincamera script is an extension of julian's PovDrone.cs so it has super smooth mouse control :p and since it'll never be disabled we don't need all the duplicate looking scripts on everything